chore: move folders around
